pyproject.toml
[project]
name = "yt-mcp-remote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.18.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "pyjwt[crypto]>=2.8.0",
    "youtube-transcript-api>=1.2.3",
    "sqlalchemy"
]
